requirements.txt
psutil
aiofiles
aiohttp
beautifulsoup4
bs4
spotipy
python-dotenv
requests
ffmpeg-python
uvloop
wget
numpy
pymediainfo
gitpython
httpx[http2]
ntgcalls
py-tgcalls
gpytranslate
gtts
Pillow
pytz
opencv-python
meval
python-dateutil
aiorun
telegraph
aiosqlite
aioshutil
speechrecognition
tgcrypto-pyrofork
qrcode
geopy
navygram
git+https://github.com/naya1503/hydrogram@dev
git+https://github.com/yt-dlp/yt-dlp@master
setproctitle